separate uncompressed jmh
